chore: update based on comments
