+++ libs/webb-ui-components/src/components/ChainsRing/index.ts
@@ -1,1 +1,3 @@
-export * from './ChainsRing';
+import { default as ChainsRing } from './ChainsRing';
+
+export default ChainsRing;
