+++ libs/webb-ui-components/src/components/ChainsRing/types.ts
@@ -1,0 +1,13 @@
+import { PropsOf } from '../../types';
+
+export interface ChainsRingProps extends PropsOf<'div'> {
+  circleContent?: React.ReactNode;
+  additionalSvgContent?: React.ReactNode;
+  chainItems: Array<ChainItem | undefined>;
+}
+
+export type ChainItem = {
+  typedChainId: number;
+  onClick?: () => void;
+  isActive?: boolean;
+};
